// frontend/src/gen/api/src/models/DisplayActivityLocationDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NewLevel
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DisplayActivityLocationDto
 */
export interface DisplayActivityLocationDto {
    /**
     * 
     * @type {string}
     * @memberof DisplayActivityLocationDto
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof DisplayActivityLocationDto
     */
    value?: number;
}

/**
 * Check if a given object implements the DisplayActivityLocationDto interface.
 */
export function instanceOfDisplayActivityLocationDto(value: object): boolean {
    return true;
}

export function DisplayActivityLocationDtoFromJSON(json: any): DisplayActivityLocationDto {
    return DisplayActivityLocationDtoFromJSONTyped(json, false);
}

export function DisplayActivityLocationDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): DisplayActivityLocationDto {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'value': json['value'] == null ? undefined : json['value'],
    };
}

export function DisplayActivityLocationDtoToJSON(value?: DisplayActivityLocationDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'name': value['name'],
        'value': value['value'],
    };
}